requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pymongo==4.5.0
werkzeug==2.3.7
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.31.0
geopy==2.3.0